+++ src/Controller/Component/AuthenticationComponent.php
@@ -79,9 +79,9 @@ class AuthenticationComponent extends BaseAuthenticationComponent
     }
 
     /**
-     * Checks whether the logged user belongs to a group.
+     * Checks whether the logged user belongs to a user group.
      *
-     * If you compare with several groups, it will check that at least one matches.
+     * If you compare with several user groups, it will check that at least one matches.
      * @param string ...$group User group
      * @return bool
      * @since 2.31.8