+++ src/Model/Table/Traits/GetPreviewsFromTextTrait.php
@@ -92,7 +92,7 @@ trait GetPreviewsFromTextTrait
     {
         $images = array_map(function (string $url): array {
             if (!is_url($url)) {
-                $url = Filesystem::instance()->makePathAbsolute($url, WWW_ROOT . 'img');
+                $url = Filesystem::makePathAbsolute($url, WWW_ROOT . 'img');
                 if (!file_exists($url)) {
                     return [];
                 }