+++ tests/TestCase/Model/Table/PostsCategoriesTableTest.php
@@ -126,6 +126,6 @@ class PostsCategoriesTableTest extends TableTestCase
         $sql = $query->sql();
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertInstanceOf(FrozenTime::class, $query->getValueBinder()->bindings()[':c1']['value']);
-        $this->assertSqlEndsWith('FROM posts_categories PostsCategories INNER JOIN posts Posts ON (Posts.active = :c0 AND Posts.created <= :c1 AND PostsCategories.id = Posts.category_id)', $sql);
+        $this->assertStringEndsWith('FROM posts_categories PostsCategories INNER JOIN posts Posts ON (Posts.active = :c0 AND Posts.created <= :c1 AND PostsCategories.id = Posts.category_id)', $sql);
     }
 }